[fix]: fix test errors and analyzejoininfo bug
